+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 **<details><summary>Fixes</summary>**
 
 * Fixed scene selection patch not matching the correct IL instructions and causing issues
+* Fixed Moons Catalogue not displaying moons that dynamically change their hidden status
 * Fixed ExtendedLevel Terminal keywords for moons with spaces not being correct
 
 </details>